Cleanup for Phpstan, PHP-CS-Fixer Upgrades
Dependabot wanted to upgrade, but changes resulted in new "errors". Fix them.

+++ src/PhpSpreadsheet/Calculation/Statistical/Deviations.php
@@ -68,7 +68,7 @@ class Deviations
         if (!is_numeric($mean)) {
             return ExcelError::DIV0();
         }
-        $stdDev = StandardDeviations::STDEV($aArgs);
+        $stdDev = (float) StandardDeviations::STDEV($aArgs);
 
         if ($stdDev > 0) {
             $count = $summer = 0;
